+++ app/__init__.py
@@ -1,6 +1,6 @@
-from flask import Flask
+from flask import Flask, render_template, request
 from flask_sqlalchemy import SQLAlchemy
-from wtforms import StringField, SubmitField, IntergerField
+
 
 app = Flask(__name__)
 app.config['SECRET_KEY'] = 'SECRET_KEY'
@@ -8,15 +8,5 @@ app.config['SQLALCHEMY_DATABASE_URI'] = "sqlite:///data.db"
 
 db = SQLAlchemy(app)
 
-class Pokedata(db.Model):
-    id = db.Column(db.Integer, primary_key =True)
-    poke_id = db.Column(db.Integer)
-    name = db.Column(db.String(30))
-    poketype = db.Column(db.String(30))
 
-    def __init__(self,poke_id,name,poketype):
-        self.poke_id = poke_id
-        self.name =  name
-        self.poketype = poketype
-
-from app import routes, models
+from app import routes, models, forms
