+++ app/templates/catchpokemon.html
@@ -5,6 +5,7 @@
 {% block body %}
 <h1>You Gotta Catch'em All!</h1>
 <div>
+    <h2> Return Home <b/> <button><a href="home">Home</a></button> </h2>
     <div>
         <div>
