+++ app/templates/insert.html
@@ -1,0 +1,31 @@
+{% extends "base.html" %}
+
+{% block title %} Poke Database {% endblock %}
+
+{% block body %}
+<h1>Add to your Pokemon Database</h1>
+<div>
+    <h2> View <b> Pokemon Database <b/> <button><a href="pokemondatabase">View Database</a></button> </h2>
+    <div class='form'>
+        <h2> Add Pokemon to Database </h2>
+        <form action=" " method = "POST">
+            {{ form.hidden_tag() }}
+            <p>
+                {{ form.poke_id.label }}    <br>
+                {{ form.poke_id (size=10)}}
+            </p>
+            <p>
+                {{ form.name.label }}    <br>
+                {{ form.name (size=30)}}
+            </p>
+            <p>
+                {{ form.poketype.label }}    <br>
+                {{ form.poketype (size=20)}}
+            </p>
+            <p>
+                {{ form.submit }}
+            </p>
+
+        </form>
+    </div>
+    {% endblock %}
