+++ app/templates/pokemondatabase.html
@@ -5,8 +5,9 @@
 {% block body %}
 <h1>Welcome to the Pokedex Database</h1>
 <div>
+    <h2> Return Home <b/> <button><a href="home">Home</a></button> </h2>
+    <h2><b> Add Pokemon <b/> <button><a href="insert">Click here</a></button> </h2>
     <div>
-        <div>
         <table>
             <tr>
                 <th>ID</th>
@@ -16,15 +17,13 @@
             </tr>
             <tr>
                 <td>01</td>
+                <td>01</td>
                 <td>Bulba</td>
                 <td>Grass</td>
             </tr>
         </table>
 
-        </div>
-
     </div>
-
 </div>
 
 {% endblock %}
